Fix bug in PromptPost & create a seperate file for PostModal
