+++ src/components/NewsFeed/NormalPost.js
@@ -11,23 +11,14 @@ import {
   Box,
   IconButton,
   Button,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
   useDisclosure,
-  Input,
   Divider,
   Center
 } from "@chakra-ui/react";
 import { FaHeart, FaComment } from "react-icons/fa";
-import Comment from './Comment';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 import { useCommentsContext } from "../../hooks/useCommentsContext";
-import { useAuthContext } from "../../hooks/useAuthContext";
+import PostModal from "./PostModal";
 
 const NormalPost = ({ post }) => {
 
@@ -37,7 +28,7 @@ const NormalPost = ({ post }) => {
   const formattedTimeStamp = formatDistanceToNow(new Date(timeStamp), { addSuffix: true })
 
   const [user, setUser] = useState(null);
-  const [newComment, setNewComment] = useState("");
+  // const [newComment, setNewComment] = useState("");
 
   const { comments, dispatch } = useCommentsContext();
 
@@ -74,32 +65,7 @@ const NormalPost = ({ post }) => {
         type: 'CLEAR_COMMENTS', 
       })
     }
-  }, [postId, userId, dispatch, isOpen]);
-
-  // Id of user currently logged in 
-  const { user: commentingUser } = useAuthContext()
-  const { _id: commentingUserId } = commentingUser.user
-  // const commentingUserId = "661f385d7bc0dc0597752644";
-
-  const handlePostComment = async () => {
-    if (!newComment.trim()) return; // Avoid posting empty comments
-
-    try {
-      const response = await axios.post(`http://localhost:4000/api/comments/${postId}`, {
-        userId: commentingUserId,
-        content: newComment
-      },{
-        headers: { "Authorization": `Bearer ${commentingUser.token}`}
-      });
-      setNewComment(""); // Clear the input field after posting the comment
-      dispatch({
-        type: 'CREATE_COMMENT',
-        payload: response.data
-      })
-    } catch (error) {
-      console.error("Error posting comment:", error);
-    }
-  };
+  }, [dispatch, isOpen]);
 
   const { avatar, username } = user || {};
 
@@ -164,51 +130,9 @@ const NormalPost = ({ post }) => {
 
       </Card>
 
-      <Modal finalFocusRef={finalRef} isOpen={isOpen} onClose={onClose} size="5xl" scrollBehavior="inside">
-            <ModalOverlay />
-            <ModalContent
-            sx={{
-                borderRadius: "30px",
-                paddingLeft: "20px",
-                paddingRight: "20px",
-            }}
-            >
-                <ModalHeader>{title}</ModalHeader>
-                <Flex flex="1" gap="5" alignItems="center" flexWrap="wrap" p={4}>
-                    {" "}
-                    {/* Added padding here */}
-                    <Avatar name={username} src={avatar} />
-                    <Box>
-                    <Text fontSize="md">{username}</Text>
-                    <Text fontSize="xs">
-                        {formattedTimeStamp}
-                    </Text>
-                    </Box>
-                </Flex>
-
-                <ModalCloseButton />
-
-                <ModalBody style={{ whiteSpace: 'pre-line' }}>
-                  {content}
-                  <Box padding={7}>
-                    <Divider w='100%' borderWidth='1px' margin={0}/>  
-                  </Box>
-                  {comments && comments.map((comment, idx) => (
-                      <Comment comment={comment} key={idx} />
-                    ))
-                  }
-                </ModalBody>
-
-
-                <ModalFooter>
-                    <Input placeholder="Your thought" value={newComment} onChange={(e) => setNewComment(e.target.value)}/>
-                    <Button colorScheme="teal" size="md" onClick={handlePostComment}>
-                    Send
-                    </Button>
-                </ModalFooter>
-            </ModalContent>
-
-        </Modal>
+      <PostModal postId={postId} isOpen={isOpen} onClose={onClose} finalRef={finalRef} 
+                title={title} username={username} avatar={avatar} formattedTimeStamp={formattedTimeStamp} 
+                content={content} comments={comments} dispatch={dispatch}/>
     </>
   );
 };
